Check to see if .gallery-image-upload exists before trying to initiate drop zone.

+++ js/zoomcomp.js
@@ -11,26 +11,27 @@
 				// autoProcessQueue: false
 			});
 
-			var dzObj = Dropzone.forElement(".gallery-image-upload");
+			if( $(".gallery-image-upload").length ){
+				var dzObj = Dropzone.forElement(".gallery-image-upload");
 
-			$('button.upload-dropzone').click(function(e){
-				e.preventDefault();
-				dzObj.processQueue();
-			});
+				$('button.upload-dropzone').click(function(e){
+					e.preventDefault();
+					dzObj.processQueue();
+				});
 
-			dzObj.
-			on('success', function(file, response){
-				dzObj.removeFile(file);
+				dzObj.
+				on('success', function(file, response){
+					dzObj.removeFile(file);
 
-				$('.existing-images').append('<li><img src="'+response.url+'"></li>');
-			}).
-			on('sending', function(file, request, formData){
-				formData.append("action", "upload_gallery_image");
-				formData.append("post_id", $("#post_ID").val());
-			})
+					$('.existing-images').append('<li><img src="'+response.url+'"></li>');
+				}).
+				on('sending', function(file, request, formData){
+					formData.append("action", "upload_gallery_image");
+					formData.append("post_id", $("#post_ID").val());
+				})
 
-
-			Dropzone.autoDiscover = false;
+				Dropzone.autoDiscover = false;
+			}
 		}
 	})
 })(jQuery);
